+++ src/blocks/text-maxi/attributes.js
@@ -4,6 +4,14 @@
 import { __ } from '@wordpress/i18n';
 
 /**
+ * Internal dependencies
+ */
+import {
+	breakpointAttributesCreator,
+	paletteAttributesCreator,
+} from '../../extensions/styles';
+
+/**
  * Imports
  */
 import * as attributesData from '../../extensions/styles/defaults/index';
@@ -45,52 +53,48 @@ const attributes = {
 		type: 'number',
 	},
 	listReversed: {
-		type: 'number',
-		default: 0,
+		type: 'boolean',
 	},
+	...breakpointAttributesCreator({
+		obj: {
+			'list-position': {
+				type: 'string',
+			},
+			'list-gap': {
+				type: 'number',
+				default: 1,
+			},
+			'list-gap-unit': {
+				type: 'string',
+				default: 'em',
+			},
+			'list-indent': {
+				type: 'number',
+			},
+			'list-indent-unit': {
+				type: 'string',
+				default: 'px',
+			},
+			'list-size': {
+				type: 'number',
+				default: 1,
+			},
+			'list-size-unit': {
+				type: 'number',
+				default: 'em',
+			},
+			'list-text-position': {
+				type: 'string',
+			},
+		},
+	}),
 	listStyle: {
 		type: 'string',
 	},
 	listStyleCustom: {
 		type: 'string',
 	},
-	listPosition: {
-		type: 'string',
-	},
-	listSVGColor: {
-		type: 'string',
-	},
-	listSize: {
-		type: 'number',
-	},
-	listSizeUnit: {
-		type: 'string',
-	},
-	// TODO: replace with future breakpointObjectCreator
-	// https://github.com/yeahcan/maxi-blocks/blob/b384ce2226e0181226817f5eda4723d1733a2f6a/src/extensions/styles/breakpointObjectCreator.js#L7
-	...(() => {
-		const response = {};
-
-		['general', 'xxl', 'xl', 'l', 'm', 's', 'xs'].forEach(breakpoint => {
-			response[`list-gap-${breakpoint}`] = {
-				type: 'number',
-				...(breakpoint === 'general' && { default: 1 }),
-			};
-			response[`list-gap-unit-${breakpoint}`] = {
-				type: 'string',
-				...(breakpoint === 'general' && { default: 'em' }),
-			};
-			response[`list-indent-${breakpoint}`] = {
-				type: 'number',
-			};
-			response[`list-indent-unit-${breakpoint}`] = {
-				type: 'string',
-				...(breakpoint === 'general' && { default: 'px' }),
-			};
-		});
-
-		return response;
-	})(),
+	...paletteAttributesCreator({ prefix: 'list-svg-', palette: 4 }),
 	...attributesData.container,
 	...attributesData.link,
 	...attributesData.textAlignment,
